Add a RepeatRule to help repeating flaky tests
A JUnit TestRule which enables to run the same JUnit test repeatedly.
This may help to identify the root cause why a flaky tests which succeed
most often does fail sometimes.

Add the RepeatRule to the test class containing the test to be repeated:

public class MyTest {
	@Rule
	public RepeatRule repeatRule = new RepeatRule();
	...
}
 
and annotate the test to be repeated with the @Repeat(n=<repetitions>)
annotation:
 
	@Test
	@Repeat(n = 100)
	public void test() {
		...
	}
 
then this test will be repeated 100 times. If any test execution fails
test repetition will be stopped.

Change-Id: I7c49ccebe1cb00bcde6b002b522d95c13fd3a35e
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>